crypto-etl/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-etl"
version = "0.1.0"
description = "Crypto ETL pipeline"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "psycopg2-binary",
    "pyyaml"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]



